Refactor: move run_ game method from game runner file to game file and fix scope issues
